method commandeline correction

+++ asmar-backend/src/main/java/com/aston/ecommerce/asmar/entity/Product.java
@@ -3,6 +3,7 @@ package com.aston.ecommerce.asmar.entity;
 import com.fasterxml.jackson.annotation.JsonBackReference;
 
 import javax.persistence.*;
+import java.math.BigDecimal;
 import java.util.List;
 
 @Entity
@@ -12,7 +13,7 @@ public class Product extends AbstractEntity{
     @Column(name="label", nullable = false)
     private String label;
     @Column(name="price", nullable = false)
-    private float price;
+    private BigDecimal price;
     @Column(name="description", nullable = false)
     private String description;
     @Column(name="history")
@@ -46,11 +47,11 @@ public class Product extends AbstractEntity{
         this.label = label;
     }
 
-    public float getPrice() {
+    public BigDecimal getPrice() {
         return price;
     }
 
-    public void setPrice(float price) {
+    public void setPrice(BigDecimal price) {
         this.price = price;
     }
